requirements.txt
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
binance-connector==1.12.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Pillow==9.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
requests==2.27.1
service-identity==21.1.0
six==1.16.0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
urllib3==1.26.9
zope.interface==5.4.0
